# Improve dedent performance.