Hey @muesli I improved dedent package performance by reimplement the logic. Shown as below.

OLD:
```hs
BenchmarkDedent-8        4960327               239 ns/op             224 B/op         10 allocs/op
BenchmarkDedent-8        4607145               242 ns/op             223 B/op         10 allocs/op
BenchmarkDedent-8        5079397               244 ns/op             224 B/op         10 allocs/op
BenchmarkDedent-8        4887057               240 ns/op             223 B/op         10 allocs/op
```

NEW:
```hs
BenchmarkDedent-8       13918354                82.9 ns/op            95 B/op          1 allocs/op
BenchmarkDedent-8       13222630                86.4 ns/op            96 B/op          1 allocs/op
BenchmarkDedent-8       13272577                87.4 ns/op            95 B/op          1 allocs/op
BenchmarkDedent-8       14463898                86.5 ns/op            95 B/op          1 allocs/op
```

But the code is little bit less intuitive than the old one. So feel free to ignore this PR.